+++ lib/e_menu/e_menu.tcl
@@ -27,7 +27,7 @@
 package require Tk
 
 namespace eval ::em {
-  variable em_version {e_menu 3.7.0}
+  variable em_version {e_menu 3.7.0a2}
   variable solo [expr {[info exist ::em::executable] || ( \
   [info exist ::argv0] && [file normalize $::argv0] eq [file normalize [info script]])} ? 1 : 0]
   variable Argv0
@@ -203,10 +203,10 @@ namespace eval ::em {
   variable hili no
   variable ls {} pk {}
   variable DF kdiff3 BF {}
-  variable PI 0 NE 0 SH 0
+  variable PI 0 NE 0
   variable th {alt} td {} g1 {} g2 {}
   variable ee {}
-  variable isbaltip yes
+  variable SH {} isbaltip yes
 }
 
 
@@ -389,7 +389,7 @@ proc ::em::focus_button {i {doit false}} {
 proc ::em::mouse_button {i} {
   focus_button $i
   set i [next_button $i]
-  if {![winfo exists .em.fr.win.fr$i.butt]} return
+  if {!$::em::isbaltip || ![winfo exists .em.fr.win.fr$i.butt]} return
   lassign [split [winfo geom .em.fr.win] +] -> x1 y1
   lassign [split [winfo geom .em.fr.win.fr$i] +x] w h x2 y2
   if {$::em::solo} {
@@ -2747,12 +2747,12 @@ proc ::em::main {args} {
 
 if {$::em::solo} {
   # theming at the app's start: th= a theme name, td= a theme directory
-  foreach ::em::TMP1 {th td} {
+  foreach ::em::TMP1 {th td SH} {
     if {[set ::em::TMP2 [lsearch -glob $::argv $::em::TMP1=*]]>-1} {
       set ::em::$::em::TMP1 [string range [lindex $::argv $::em::TMP2] 3 end]
     }
   }
-  set ::em::isbaltip [expr {{SH=1} ni $::argv}]
+  set ::em::isbaltip [expr {$::em::SH eq {}}]
   if {$::em::th eq {} || $::em::td eq {}} {
     ::apave::initWM -isbaltip $::em::isbaltip
   } else {